week05/requirements.txt
langchain==1.0.2
langchain-google-genai==3.0.0
python-dotenv==1.2.1
